Second Example Output for the pathfindR Enrichment Workflow

The data frame containing the results of pathfindR's active-subnetwork-oriented enrichment workflow performed on the rheumatoid arthritis dataset GSE84074 https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE84074. Analysis via run_pathfindR was performed using the default settings. Generated on Feb 9, 2020.

RA_comparison_output

Format

A data frame with 35 rows and 8 columns:

ID

ID of the enriched term

Term_Description

Description of the enriched term

Fold_Enrichment

Fold enrichment value for the enriched term

occurrence

the number of iterations that the given term was found to enriched over all iterations

lowest_p

the lowest adjusted-p value of the given term over all iterations

highest_p

the highest adjusted-p value of the given term over all iterations

Up_regulated

the up-regulated genes in the input involved in the given term, comma-separated

Down_regulated

the down-regulated genes in the input involved in the given term, comma-separated

See also