Retrieve Organism-specific PIN data

Retrieve Organism-specific PIN data

get_pin_file(source = "BioGRID", org = "Homo_sapiens", path2pin, ...)

Arguments

source

As of this version, this function is implemented to get data from "BioGRID" only. This argument (and this wrapper function) was implemented for future utility

org

organism name. BioGRID naming requires underscores for spaces so "Homo sapiens" becomes "Homo_sapiens", "Mus musculus" becomes "Mus_musculus" etc. See https://wiki.thebiogrid.org/doku.php/statistics for a full list of available organisms (default = "Homo_sapiens")

path2pin

the path of the file to save the PIN data. By default, the PIN data is saved in a temporary file

...

additional arguments for get_biogrid_pin

Value

the path of the file in which the PIN data was saved. If path2pin was not supplied by the user, the PIN data is saved in a temporary file

Examples

if (FALSE) {
pin_path <- get_pin_file()
}