Process Input

Process Input

input_processing(
  input,
  p_val_threshold = 0.05,
  pin_name_path = "Biogrid",
  convert2alias = TRUE
)

Arguments

input
the input data that pathfindR uses. The input must be a data frame with three columns:

Gene Symbol (Gene Symbol)

Change value, e.g. log(fold change) (OPTIONAL)

p value, e.g. adjusted p value associated with differential expression

p_val_threshold
the p value threshold to use when filtering the input data frame. Must a numeric value between 0 and 1. (default = 0.05)

pin_name_path
Name of the chosen PIN or path/to/PIN.sif. If PIN name, must be one of c("Biogrid", "STRING", "GeneMania", "IntAct", "KEGG", "mmu_STRING"). If path/to/PIN.sif, the file must comply with the PIN specifications. (Default = "Biogrid")

convert2alias
boolean to indicate whether or not to convert gene symbols in the input that are not found in the PIN to an alias symbol found in the PIN (default = TRUE) IMPORTANT NOTE: the conversion uses human gene symbols/alias symbols.

Value

This function first filters the input so that all p values are less than or equal to the threshold. Next, gene symbols that are not found in the PIN are identified. If aliases of these gene symbols are found in the PIN, the symbols are converted to the corresponding aliases. The resulting data frame containing the original gene symbols, the updated symbols, change values and p values is then returned.

See also

Examples

# \dontshow{
input_processing(RA_input[1:20, ], 0.05, "KEGG")

#> Number of genes provided in input: 20

#> Number of genes in input after p-value filtering: 20

#> 

#> Could not find any interactions for 8 (40%) genes in the PIN

#> Final number of genes in input: 12

#>    old_GENE     GENE     CHANGE      P_VALUE
#> 1   FAM110A      F10 -0.6939359 3.408699e-06
#> 3    S100A8      MIF  1.5448338 3.466386e-05
#> 4    S100A9   S100A9  1.0280904 2.262617e-04
#> 6  ARHGAP17 ARHGAP17 -0.6919330 2.708081e-04
#> 7     NUP62    NUP62 -0.8216320 2.708081e-04
#> 8     MYL6B    MYL6B  0.2790078 4.006813e-04
#> 11     KLF2     KLF2 -0.4734862 4.512990e-04
#> 12    RUNX3    RUNX3 -0.9341854 4.512990e-04
#> 14      TXN      TXN  1.0707134 5.608094e-04
#> 18      GNE      GNE -0.4718260 6.330627e-04
#> 19    GDPD3    GDPD3  0.5655699 6.330627e-04
#> 20    NUP93    NUP93 -0.4655986 6.330627e-04

input_processing(RA_input[1:20, ], 0.05, "KEGG", convert2alias = FALSE)

#> Number of genes provided in input: 20

#> Number of genes in input after p-value filtering: 20

#> Could not find any interactions for 11 (55%) genes in the PIN

#> Final number of genes in input: 9

#>    old_GENE     GENE     CHANGE      P_VALUE
#> 6  ARHGAP17 ARHGAP17 -0.6919330 0.0002708081
#> 7     NUP62    NUP62 -0.8216320 0.0002708081
#> 8     MYL6B    MYL6B  0.2790078 0.0004006813
#> 11     KLF2     KLF2 -0.4734862 0.0004512990
#> 12    RUNX3    RUNX3 -0.9341854 0.0004512990
#> 14      TXN      TXN  1.0707134 0.0005608094
#> 18      GNE      GNE -0.4718260 0.0006330627
#> 19    GDPD3    GDPD3  0.5655699 0.0006330627
#> 20    NUP93    NUP93 -0.4655986 0.0006330627

# }
if (FALSE) {
input_processing(RA_input, 0.05, "KEGG")
}