Summarize Enrichment Results

Summarize Enrichment Results

summarize_enrichment_results(enrichment_res, list_active_snw_genes = FALSE)

Arguments

enrichment_res

a dataframe of combined enrichment results. Columns are:

ID

ID of the enriched term

Term_Description

Description of the enriched term

Fold_Enrichment

Fold enrichment value for the enriched term

p_value

p value of enrichment

adj_p

adjusted p value of enrichment

non_Signif_Snw_Genes (OPTIONAL)

the non-significant active subnetwork genes, comma-separated

list_active_snw_genes

boolean value indicating whether or not to report the non-significant active subnetwork genes for the active subnetwork which was enriched for the given term with the lowest p value (default = FALSE)

Value

a dataframe of summarized enrichment results (over multiple iterations). Columns are:

ID

ID of the enriched term

Term_Description

Description of the enriched term

Fold_Enrichment

Fold enrichment value for the enriched term

occurrence

the number of iterations that the given term was found to enriched over all iterations

support

the median support (proportion of active subnetworks leading to enrichment within an iteration) over all iterations

lowest_p

the lowest adjusted-p value of the given term over all iterations

highest_p

the highest adjusted-p value of the given term over all iterations

non_Signif_Snw_Genes (OPTIONAL)

the non-significant active subnetwork genes, comma-separated

Examples

if (FALSE) {
summarize_enrichment_results(enrichment_res)
}