Visualize Human KEGG Pathways

Visualize Human KEGG Pathways

visualize_hsa_KEGG(
  hsa_kegg_ids,
  input_processed,
  scale_vals = TRUE,
  node_cols = NULL,
  quiet = TRUE,
  key_gravity = "northeast",
  logo_gravity = "southeast"
)

Arguments

hsa_kegg_ids

hsa KEGG ids of pathways to be colored and visualized

input_processed

input data processed via input_processing

scale_vals

should change values be scaled? (default = TRUE)

node_cols

low, middle and high color values for coloring the pathway nodes (default = NULL). If node_cols=NULL, the low, middle and high color are set as "green", "gray" and "red". If all change values are 1e6 (in case no changes are supplied, this dummy value is assigned by input_processing), only one color ("#F38F18" if NULL) is used.

quiet

If TRUE, suppress status messages (if any), and the progress bar while downloading file(s)

key_gravity

gravity value (character) for the color key legend placement (see gravity_types)

logo_gravity

gravity value (character) for the logo placement (see gravity_types)

Value

Creates colored visualizations of the enriched human KEGG pathways and saves them in the folder "term_visualizations" under the current working directory.

See also

Examples

if (FALSE) {
visualize_hsa_KEGG(hsa_kegg_ids, input_processed)
}